+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] - 2023-03-31
+
+* Updates slic-stack.site.yml for full site testing.
+* Adds check for `site` target to refrence `slic_wp_dir()` when running commands.
+
 ## [1.3.1] - 2023-03-31
 
 * Change - Default to Composer version 2.