+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3] - 2023-04-03
+
+* Fix - broken CRLF line breaks via slic real-time commands due to docker compose v2.2.x being used [Docker compose issue](https://github.com/docker/compose/issues/8833#issuecomment-953023240).
+
 ## [1.3.2] - 2023-04-03
 
 * Updates slic-stack.site.yml for full site testing.